+++ packages/core/src/main/conversation/ConversationService/ConversationService.types.ts
@@ -125,3 +125,18 @@ export type SendProteusMessageParams<T> = SendCommonParams<T> &
       wasSent: boolean,
     ) => void | boolean | Promise<boolean>;
   };
+
+export type SendMlsMessageParams<T> = SendCommonParams<T> & {
+  /**
+   * The groupId of the conversation to send the message to (Needed only for MLS)
+   */
+  groupId: string;
+};
+
+export type QualifiedUsers = QualifiedId & {skipOwn?: string};
+
+export type AddUsersParams = {
+  conversationId: QualifiedId;
+  qualifiedUserIds: QualifiedId[];
+  groupId?: string;
+};
